requirements.txt
colorama==0.4.6
openai==1.57.0
python-dotenv==1.0.0
retry==0.9.2
ruamel.yaml==0.17.32
tiktoken==0.8.0
tqdm==4.65.0
pillow==10.4.0
PyQt5>=5.15.0
regex==2025.9.1
rich==14.1.0
click==8.2.1
construct==2.10.70
pandas==2.3.2
